SignersProcessor
