+++ services/horizon/internal/db2/history/mock_q_signers.go
@@ -43,3 +43,8 @@ func (m *MockQSigners) RemoveAccountSigner(account, signer string) (int64, error
 	a := m.Called(account, signer)
 	return a.Get(0).(int64), a.Error(1)
 }
+
+func (m *MockQSigners) CountAccounts() (int, error) {
+	a := m.Called()
+	return a.Get(0).(int), a.Error(1)
+}
