+++ services/horizon/internal/expingest/processors/signer_processor_test.go
@@ -1,0 +1,600 @@
+package processors
+
+import (
+	"testing"
+
+	ingesterrors "github.com/stellar/go/exp/ingest/errors"
+	"github.com/stellar/go/exp/ingest/io"
+	"github.com/stellar/go/services/horizon/internal/db2/history"
+	"github.com/stellar/go/support/errors"
+	"github.com/stellar/go/xdr"
+	"github.com/stretchr/testify/suite"
+)
+
+func TestAccountsSignerProcessorTestSuiteState(t *testing.T) {
+	suite.Run(t, new(AccountsSignerProcessorTestSuiteState))
+}
+
+type AccountsSignerProcessorTestSuiteState struct {
+	suite.Suite
+	processor              *SignersProcessor
+	mockQ                  *history.MockQSigners
+	mockBatchInsertBuilder *history.MockAccountSignersBatchInsertBuilder
+}
+
+func (s *AccountsSignerProcessorTestSuiteState) SetupTest() {
+	s.mockQ = &history.MockQSigners{}
+	s.mockBatchInsertBuilder = &history.MockAccountSignersBatchInsertBuilder{}
+
+	s.mockQ.
+		On("NewAccountSignersBatchInsertBuilder", maxBatchSize).
+		Return(s.mockBatchInsertBuilder).Once()
+
+	s.processor = &SignersProcessor{
+		SignersQ: s.mockQ,
+	}
+}
+
+func (s *AccountsSignerProcessorTestSuiteState) TearDownTest() {
+	s.mockBatchInsertBuilder.On("Exec").Return(nil).Once()
+	s.Assert().NoError(s.processor.Commit())
+
+	s.mockQ.AssertExpectations(s.T())
+	s.mockBatchInsertBuilder.AssertExpectations(s.T())
+}
+
+func (s *AccountsSignerProcessorTestSuiteState) TestNoEntries() {
+	s.mockQ.On("CountAccounts").Return(0, nil).Once()
+
+	header := xdr.LedgerHeader{
+		LedgerSeq: xdr.Uint32(63),
+	}
+	err := s.processor.Init(header)
+	s.Assert().NoError(err)
+
+	// Nothing processed, assertions in TearDownTest.
+}
+
+func (s *AccountsSignerProcessorTestSuiteState) TestCreatesSigners() {
+	s.mockQ.On("CountAccounts").Return(0, nil).Once()
+
+	header := xdr.LedgerHeader{
+		LedgerSeq: xdr.Uint32(63),
+	}
+	err := s.processor.Init(header)
+	s.Assert().NoError(err)
+
+	s.mockBatchInsertBuilder.
+		On("Add", history.AccountSigner{
+			Account: "GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			Signer:  "GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			Weight:  int32(1),
+		}).Return(nil).Once()
+
+	err = s.processor.ProcessChange(io.Change{
+		Type: xdr.LedgerEntryTypeAccount,
+		Pre:  nil,
+		Post: &xdr.LedgerEntry{
+			Data: xdr.LedgerEntryData{
+				Type: xdr.LedgerEntryTypeAccount,
+				Account: &xdr.AccountEntry{
+					AccountId:  xdr.MustAddress("GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML"),
+					Thresholds: [4]byte{1, 1, 1, 1},
+				},
+			},
+		},
+	})
+	s.Assert().NoError(err)
+
+	s.mockBatchInsertBuilder.
+		On("Add", history.AccountSigner{
+			Account: "GCCCU34WDY2RATQTOOQKY6SZWU6J5DONY42SWGW2CIXGW4LICAGNRZKX",
+			Signer:  "GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			Weight:  int32(10),
+		}).Return(nil).Once()
+
+	err = s.processor.ProcessChange(io.Change{
+		Type: xdr.LedgerEntryTypeAccount,
+		Pre:  nil,
+		Post: &xdr.LedgerEntry{
+			Data: xdr.LedgerEntryData{
+				Type: xdr.LedgerEntryTypeAccount,
+				Account: &xdr.AccountEntry{
+					AccountId: xdr.MustAddress("GCCCU34WDY2RATQTOOQKY6SZWU6J5DONY42SWGW2CIXGW4LICAGNRZKX"),
+					Signers: []xdr.Signer{
+						xdr.Signer{
+							Key:    xdr.MustSigner("GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML"),
+							Weight: 10,
+						},
+					},
+				},
+			},
+		},
+	})
+	s.Assert().NoError(err)
+
+}
+
+func TestAccountsSignerProcessorTestSuiteLedger(t *testing.T) {
+	suite.Run(t, new(AccountsSignerProcessorTestSuiteLedger))
+}
+
+type AccountsSignerProcessorTestSuiteLedger struct {
+	suite.Suite
+	processor *SignersProcessor
+	mockQ     *history.MockQSigners
+}
+
+func (s *AccountsSignerProcessorTestSuiteLedger) SetupTest() {
+	s.mockQ = &history.MockQSigners{}
+	s.mockQ.
+		On("NewAccountSignersBatchInsertBuilder", maxBatchSize).
+		Return(&history.MockAccountSignersBatchInsertBuilder{}).Once()
+
+	s.processor = &SignersProcessor{
+		SignersQ: s.mockQ,
+	}
+
+	// No insertOnlyMode
+	s.mockQ.On("CountAccounts").Return(100, nil).Once()
+
+	header := xdr.LedgerHeader{
+		LedgerSeq: xdr.Uint32(63),
+	}
+	err := s.processor.Init(header)
+	s.Assert().NoError(err)
+}
+
+func (s *AccountsSignerProcessorTestSuiteLedger) TearDownTest() {
+	s.mockQ.AssertExpectations(s.T())
+}
+
+func (s *AccountsSignerProcessorTestSuiteLedger) TestNoTransactions() {
+	// Nothing processed, assertions in TearDownTest.
+	s.Assert().NoError(s.processor.Commit())
+}
+
+func (s *AccountsSignerProcessorTestSuiteLedger) TestNewAccount() {
+	s.mockQ.
+		On(
+			"CreateAccountSigner",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			int32(1),
+		).
+		Return(int64(1), nil).Once()
+
+	err := s.processor.ProcessChange(io.Change{
+		Type: xdr.LedgerEntryTypeAccount,
+		Pre:  nil,
+		Post: &xdr.LedgerEntry{
+			Data: xdr.LedgerEntryData{
+				Type: xdr.LedgerEntryTypeAccount,
+				Account: &xdr.AccountEntry{
+					AccountId:  xdr.MustAddress("GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML"),
+					Thresholds: [4]byte{1, 1, 1, 1},
+				},
+			},
+		},
+	})
+	s.Assert().NoError(err)
+	s.Assert().NoError(s.processor.Commit())
+}
+
+func (s *AccountsSignerProcessorTestSuiteLedger) TestNewSigner() {
+	// Remove old signer
+	s.mockQ.
+		On(
+			"RemoveAccountSigner",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			"GCBBDQLCTNASZJ3MTKAOYEOWRGSHDFAJVI7VPZUOP7KXNHYR3HP2BUKV",
+		).
+		Return(int64(1), nil).Once()
+
+	// Create new and old signer
+	s.mockQ.
+		On(
+			"CreateAccountSigner",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			"GCBBDQLCTNASZJ3MTKAOYEOWRGSHDFAJVI7VPZUOP7KXNHYR3HP2BUKV",
+			int32(10),
+		).
+		Return(int64(1), nil).Once()
+
+	s.mockQ.
+		On(
+			"CreateAccountSigner",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			"GCAHY6JSXQFKWKP6R7U5JPXDVNV4DJWOWRFLY3Y6YPBF64QRL4BPFDNS",
+			int32(15),
+		).
+		Return(int64(1), nil).Once()
+
+	err := s.processor.ProcessChange(io.Change{
+		Type: xdr.LedgerEntryTypeAccount,
+		Pre: &xdr.LedgerEntry{
+			Data: xdr.LedgerEntryData{
+				Type: xdr.LedgerEntryTypeAccount,
+				Account: &xdr.AccountEntry{
+					AccountId: xdr.MustAddress("GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML"),
+					Signers: []xdr.Signer{
+						xdr.Signer{
+							Key:    xdr.MustSigner("GCBBDQLCTNASZJ3MTKAOYEOWRGSHDFAJVI7VPZUOP7KXNHYR3HP2BUKV"),
+							Weight: 10,
+						},
+					},
+				},
+			},
+		},
+		Post: &xdr.LedgerEntry{
+			Data: xdr.LedgerEntryData{
+				Type: xdr.LedgerEntryTypeAccount,
+				Account: &xdr.AccountEntry{
+					AccountId: xdr.MustAddress("GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML"),
+					Signers: []xdr.Signer{
+						xdr.Signer{
+							Key:    xdr.MustSigner("GCBBDQLCTNASZJ3MTKAOYEOWRGSHDFAJVI7VPZUOP7KXNHYR3HP2BUKV"),
+							Weight: 10,
+						},
+						xdr.Signer{
+							Key:    xdr.MustSigner("GCAHY6JSXQFKWKP6R7U5JPXDVNV4DJWOWRFLY3Y6YPBF64QRL4BPFDNS"),
+							Weight: 15,
+						},
+					},
+				},
+			},
+		},
+	})
+	s.Assert().NoError(err)
+	s.Assert().NoError(s.processor.Commit())
+}
+
+func (s *AccountsSignerProcessorTestSuiteLedger) TestSignerRemoved() {
+	// Remove old signers
+	s.mockQ.
+		On(
+			"RemoveAccountSigner",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			"GCBBDQLCTNASZJ3MTKAOYEOWRGSHDFAJVI7VPZUOP7KXNHYR3HP2BUKV",
+		).
+		Return(int64(1), nil).Once()
+
+	s.mockQ.
+		On(
+			"RemoveAccountSigner",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			"GCAHY6JSXQFKWKP6R7U5JPXDVNV4DJWOWRFLY3Y6YPBF64QRL4BPFDNS",
+		).
+		Return(int64(1), nil).Once()
+
+	// Create new signer
+	s.mockQ.
+		On(
+			"CreateAccountSigner",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			"GCAHY6JSXQFKWKP6R7U5JPXDVNV4DJWOWRFLY3Y6YPBF64QRL4BPFDNS",
+			int32(15),
+		).
+		Return(int64(1), nil).Once()
+
+	err := s.processor.ProcessChange(io.Change{
+		Type: xdr.LedgerEntryTypeAccount,
+		Pre: &xdr.LedgerEntry{
+			Data: xdr.LedgerEntryData{
+				Type: xdr.LedgerEntryTypeAccount,
+				Account: &xdr.AccountEntry{
+					AccountId: xdr.MustAddress("GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML"),
+					Signers: []xdr.Signer{
+						xdr.Signer{
+							Key:    xdr.MustSigner("GCBBDQLCTNASZJ3MTKAOYEOWRGSHDFAJVI7VPZUOP7KXNHYR3HP2BUKV"),
+							Weight: 10,
+						},
+						xdr.Signer{
+							Key:    xdr.MustSigner("GCAHY6JSXQFKWKP6R7U5JPXDVNV4DJWOWRFLY3Y6YPBF64QRL4BPFDNS"),
+							Weight: 15,
+						},
+					},
+				},
+			},
+		},
+		Post: &xdr.LedgerEntry{
+			Data: xdr.LedgerEntryData{
+				Type: xdr.LedgerEntryTypeAccount,
+				Account: &xdr.AccountEntry{
+					AccountId: xdr.MustAddress("GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML"),
+					Signers: []xdr.Signer{
+						xdr.Signer{
+							Key:    xdr.MustSigner("GCAHY6JSXQFKWKP6R7U5JPXDVNV4DJWOWRFLY3Y6YPBF64QRL4BPFDNS"),
+							Weight: 15,
+						},
+					},
+				},
+			},
+		},
+	})
+	s.Assert().NoError(err)
+	s.Assert().NoError(s.processor.Commit())
+}
+
+// TestSignerPreAuthTxRemovedTxFailed tests if removing preauthorized transaction
+// signer works even when tx failed.
+func (s *AccountsSignerProcessorTestSuiteLedger) TestSignerPreAuthTxRemovedTxFailed() {
+	// Remove old signers
+	s.mockQ.
+		On(
+			"RemoveAccountSigner",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			"GCBBDQLCTNASZJ3MTKAOYEOWRGSHDFAJVI7VPZUOP7KXNHYR3HP2BUKV",
+		).
+		Return(int64(1), nil).Once()
+
+	s.mockQ.
+		On(
+			"RemoveAccountSigner",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			"TBU2RRGLXH3E5CQHTD3ODLDF2BWDCYUSSBLLZ5GNW7JXHDIYKXZWHXL7",
+		).
+		Return(int64(1), nil).Once()
+
+	// Create new signer
+	s.mockQ.
+		On(
+			"CreateAccountSigner",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			"GCBBDQLCTNASZJ3MTKAOYEOWRGSHDFAJVI7VPZUOP7KXNHYR3HP2BUKV",
+			int32(10),
+		).
+		Return(int64(1), nil).Once()
+
+	err := s.processor.ProcessChange(io.Change{
+		Type: xdr.LedgerEntryTypeAccount,
+		Pre: &xdr.LedgerEntry{
+			Data: xdr.LedgerEntryData{
+				Type: xdr.LedgerEntryTypeAccount,
+				Account: &xdr.AccountEntry{
+					AccountId: xdr.MustAddress("GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML"),
+					Signers: []xdr.Signer{
+						xdr.Signer{
+							Key:    xdr.MustSigner("GCBBDQLCTNASZJ3MTKAOYEOWRGSHDFAJVI7VPZUOP7KXNHYR3HP2BUKV"),
+							Weight: 10,
+						},
+						xdr.Signer{
+							Key:    xdr.MustSigner("TBU2RRGLXH3E5CQHTD3ODLDF2BWDCYUSSBLLZ5GNW7JXHDIYKXZWHXL7"),
+							Weight: 15,
+						},
+					},
+				},
+			},
+		},
+		Post: &xdr.LedgerEntry{
+			Data: xdr.LedgerEntryData{
+				Type: xdr.LedgerEntryTypeAccount,
+				Account: &xdr.AccountEntry{
+					AccountId: xdr.MustAddress("GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML"),
+					Signers: []xdr.Signer{
+						xdr.Signer{
+							Key:    xdr.MustSigner("GCBBDQLCTNASZJ3MTKAOYEOWRGSHDFAJVI7VPZUOP7KXNHYR3HP2BUKV"),
+							Weight: 10,
+						},
+					},
+				},
+			},
+		},
+	})
+	s.Assert().NoError(err)
+	s.Assert().NoError(s.processor.Commit())
+}
+
+func (s *AccountsSignerProcessorTestSuiteLedger) TestRemoveAccount() {
+	s.mockQ.
+		On(
+			"RemoveAccountSigner",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+		).
+		Return(int64(1), nil).Once()
+
+	err := s.processor.ProcessChange(io.Change{
+		Type: xdr.LedgerEntryTypeAccount,
+		Pre: &xdr.LedgerEntry{
+			Data: xdr.LedgerEntryData{
+				Type: xdr.LedgerEntryTypeAccount,
+				Account: &xdr.AccountEntry{
+					AccountId:  xdr.MustAddress("GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML"),
+					Thresholds: [4]byte{1, 1, 1, 1},
+				},
+			},
+		},
+		Post: nil,
+	})
+	s.Assert().NoError(err)
+	s.Assert().NoError(s.processor.Commit())
+}
+
+func (s *AccountsSignerProcessorTestSuiteLedger) TestNewAccountNoRowsAffected() {
+	s.mockQ.
+		On(
+			"CreateAccountSigner",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			int32(1),
+		).
+		Return(int64(0), nil).Once()
+
+	err := s.processor.ProcessChange(io.Change{
+		Type: xdr.LedgerEntryTypeAccount,
+		Pre:  nil,
+		Post: &xdr.LedgerEntry{
+			Data: xdr.LedgerEntryData{
+				Type: xdr.LedgerEntryTypeAccount,
+				Account: &xdr.AccountEntry{
+					AccountId:  xdr.MustAddress("GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML"),
+					Thresholds: [4]byte{1, 1, 1, 1},
+				},
+			},
+		},
+	})
+	s.Assert().NoError(err)
+
+	err = s.processor.Commit()
+	s.Assert().Error(err)
+	s.Assert().IsType(ingesterrors.StateError{}, errors.Cause(err))
+	s.Assert().EqualError(
+		err,
+		"0 rows affected when inserting "+
+			"account=GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML "+
+			"signer=GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML to database",
+	)
+}
+
+func (s *AccountsSignerProcessorTestSuiteLedger) TestRemoveAccountNoRowsAffected() {
+	s.mockQ.
+		On(
+			"RemoveAccountSigner",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+		).
+		Return(int64(0), nil).Once()
+
+	err := s.processor.ProcessChange(io.Change{
+		Type: xdr.LedgerEntryTypeAccount,
+		Pre: &xdr.LedgerEntry{
+			Data: xdr.LedgerEntryData{
+				Type: xdr.LedgerEntryTypeAccount,
+				Account: &xdr.AccountEntry{
+					AccountId:  xdr.MustAddress("GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML"),
+					Thresholds: [4]byte{1, 1, 1, 1},
+				},
+			},
+		},
+		Post: nil,
+	})
+	s.Assert().NoError(err)
+
+	err = s.processor.Commit()
+	s.Assert().Error(err)
+	s.Assert().IsType(ingesterrors.StateError{}, errors.Cause(err))
+	s.Assert().EqualError(
+		err,
+		"Expected "+
+			"account=GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML "+
+			"signer=GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML in database but not found when removing "+
+			"(rows affected = 0)",
+	)
+}
+
+func (s *AccountsSignerProcessorTestSuiteLedger) TestProcessUpgradeChange() {
+	// Remove old signer
+	s.mockQ.
+		On(
+			"RemoveAccountSigner",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			"GCBBDQLCTNASZJ3MTKAOYEOWRGSHDFAJVI7VPZUOP7KXNHYR3HP2BUKV",
+		).
+		Return(int64(1), nil).Once()
+
+	// Create new and old (updated) signer
+	s.mockQ.
+		On(
+			"CreateAccountSigner",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			"GCBBDQLCTNASZJ3MTKAOYEOWRGSHDFAJVI7VPZUOP7KXNHYR3HP2BUKV",
+			int32(12),
+		).
+		Return(int64(1), nil).Once()
+
+	s.mockQ.
+		On(
+			"CreateAccountSigner",
+			"GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML",
+			"GCAHY6JSXQFKWKP6R7U5JPXDVNV4DJWOWRFLY3Y6YPBF64QRL4BPFDNS",
+			int32(15),
+		).
+		Return(int64(1), nil).Once()
+
+	err := s.processor.ProcessChange(io.Change{
+		Type: xdr.LedgerEntryTypeAccount,
+		Pre: &xdr.LedgerEntry{
+			Data: xdr.LedgerEntryData{
+				Type: xdr.LedgerEntryTypeAccount,
+				Account: &xdr.AccountEntry{
+					AccountId: xdr.MustAddress("GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML"),
+					Signers: []xdr.Signer{
+						xdr.Signer{
+							Key:    xdr.MustSigner("GCBBDQLCTNASZJ3MTKAOYEOWRGSHDFAJVI7VPZUOP7KXNHYR3HP2BUKV"),
+							Weight: 10,
+						},
+					},
+				},
+			},
+		},
+		Post: &xdr.LedgerEntry{
+			Data: xdr.LedgerEntryData{
+				Type: xdr.LedgerEntryTypeAccount,
+				Account: &xdr.AccountEntry{
+					AccountId: xdr.MustAddress("GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML"),
+					Signers: []xdr.Signer{
+						xdr.Signer{
+							Key:    xdr.MustSigner("GCBBDQLCTNASZJ3MTKAOYEOWRGSHDFAJVI7VPZUOP7KXNHYR3HP2BUKV"),
+							Weight: 10,
+						},
+						xdr.Signer{
+							Key:    xdr.MustSigner("GCAHY6JSXQFKWKP6R7U5JPXDVNV4DJWOWRFLY3Y6YPBF64QRL4BPFDNS"),
+							Weight: 15,
+						},
+					},
+				},
+			},
+		},
+	})
+	s.Assert().NoError(err)
+
+	err = s.processor.ProcessChange(io.Change{
+		Type: xdr.LedgerEntryTypeAccount,
+		Pre: &xdr.LedgerEntry{
+			LastModifiedLedgerSeq: 1000,
+			Data: xdr.LedgerEntryData{
+				Type: xdr.LedgerEntryTypeAccount,
+				Account: &xdr.AccountEntry{
+					AccountId: xdr.MustAddress("GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML"),
+					Signers: []xdr.Signer{
+						xdr.Signer{
+							Key:    xdr.MustSigner("GCBBDQLCTNASZJ3MTKAOYEOWRGSHDFAJVI7VPZUOP7KXNHYR3HP2BUKV"),
+							Weight: 10,
+						},
+					},
+				},
+			},
+		},
+		Post: &xdr.LedgerEntry{
+			LastModifiedLedgerSeq: 1001,
+			Data: xdr.LedgerEntryData{
+				Type: xdr.LedgerEntryTypeAccount,
+				Account: &xdr.AccountEntry{
+					AccountId: xdr.MustAddress("GC3C4AKRBQLHOJ45U4XG35ESVWRDECWO5XLDGYADO6DPR3L7KIDVUMML"),
+					Signers: []xdr.Signer{
+						xdr.Signer{
+							Key:    xdr.MustSigner("GCBBDQLCTNASZJ3MTKAOYEOWRGSHDFAJVI7VPZUOP7KXNHYR3HP2BUKV"),
+							Weight: 12,
+						},
+						xdr.Signer{
+							Key:    xdr.MustSigner("GCAHY6JSXQFKWKP6R7U5JPXDVNV4DJWOWRFLY3Y6YPBF64QRL4BPFDNS"),
+							Weight: 15,
+						},
+					},
+				},
+			},
+		},
+	})
+	s.Assert().NoError(err)
+	s.Assert().NoError(s.processor.Commit())
+}
+
+func createTransactionMeta(opMeta []xdr.OperationMeta) xdr.TransactionMeta {
+	return xdr.TransactionMeta{
+		V: 1,
+		V1: &xdr.TransactionMetaV1{
+			Operations: opMeta,
+		},
+	}
+}
